+++ CHANGES.md
@@ -297,5 +297,7 @@
 - Portable release: bundled with [python-3.13.7-embed-win32](https://www.python.org/ftp/python/3.13.7/python-3.13.7-embed-win32.zip) package and necessary modules
 
 ### Pre-release (work-in-progress)
+- Add support for Miracle-Ear Harmony II
 - ReSound/Beltone/Interton: display full version numbers
+- Quick Scan: fix `config.ini` missing bug
 - Document update